moved the files to seperate directories inside UI dir
