+++ business_banking/lib/features/budget/ui/chart/view_chart_widget.dart
@@ -1,14 +1,14 @@
 import 'package:business_banking/features/budget/bloc/budget_bloc.dart';
-import 'package:business_banking/features/budget/ui/budget_presenter.dart';
+import 'package:business_banking/features/budget/ui/chart/view_chart_presenter.dart';
 import 'package:clean_framework/clean_framework.dart';
 import 'package:flutter/material.dart';
 
-class BudgetWidget extends StatelessWidget {
+class ViewChartWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return BlocProvider<BudgetBloc>(
       create: (_) => BudgetBloc(),
-      child: BudgetPresenter(),
+      child: ViewChartPresenter(),
     );
   }
 }
