+++ business_banking/lib/features/budget/ui/first_card/budget_widget.dart
@@ -1,14 +1,14 @@
 import 'package:business_banking/features/budget/bloc/budget_bloc.dart';
-import 'package:business_banking/features/budget/ui/list_transactions_presenter.dart';
+import 'package:business_banking/features/budget/ui/first_card/budget_presenter.dart';
 import 'package:clean_framework/clean_framework.dart';
 import 'package:flutter/material.dart';
 
-class ListTransactionsWidget extends StatelessWidget {
+class BudgetWidget extends StatelessWidget {
   @override
-  Widget build(BuildContext build) {
+  Widget build(BuildContext context) {
     return BlocProvider<BudgetBloc>(
       create: (_) => BudgetBloc(),
-      child: ListTransactionsPresenter(),
+      child: BudgetPresenter(),
     );
   }
 }
