+++ business_banking/lib/features/budget/ui/transactions_list/list_transactions_presenter.dart
@@ -1,6 +1,6 @@
 import 'package:business_banking/features/budget/bloc/budget_bloc.dart';
 import 'package:business_banking/features/budget/model/list_transactions_view_model.dart';
-import 'package:business_banking/features/budget/ui/list_transactions_screen.dart';
+import 'package:business_banking/features/budget/ui/transactions_list/list_transactions_screen.dart';
 import 'package:clean_framework/clean_framework.dart';
 import 'package:flutter/material.dart';
